+++ lmcache/integration/vllm/vllm_v1_adapter.py
@@ -77,76 +77,6 @@ def get_zmq_rpc_path_lmcache(
     logger.debug("Base URL: %s, RPC Port: %s", base_url, rpc_port)
     return f"ipc://{base_url}/lmcache_rpc_port_{rpc_port}"
 
-
-# TODO: move this to LMCache so that we can gracefully close it
-class LMCacheLookupClient:
-    def __init__(self, role: KVConnectorRole, is_tp: bool, vllm_config: "VllmConfig"):
-        self.encoder = MsgpackEncoder()
-        self.ctx = zmq.Context()  # type: ignore[attr-defined]
-        socket_path = get_zmq_rpc_path_lmcache(role, is_tp, vllm_config)
-        self.socket = make_zmq_socket(
-            self.ctx,
-            socket_path,
-            zmq.REQ,  # type: ignore[attr-defined]
-            bind=False,
-        )
-
-    def lookup(self, token_ids: torch.Tensor, user: Optional[str] = "") -> int:
-        self.socket.send_multipart([pickle.dumps(user), pickle.dumps(token_ids)], copy=False)
-        resp = self.socket.recv()
-        result = int.from_bytes(resp, "big")
-        return result
-
-    def close(self):
-        self.socket.close(linger=0)
-
-
-class LMCacheLookupServer:
-    def __init__(
-        self,
-        lmcache_engine: LMCacheEngine,
-        role: KVConnectorRole,
-        is_tp: bool,
-        vllm_config: "VllmConfig",
-    ):
-        self.decoder = MsgpackDecoder(torch.Tensor)
-        self.ctx = zmq.Context()  # type: ignore[attr-defined]
-        socket_path = get_zmq_rpc_path_lmcache(role, is_tp, vllm_config)
-        self.socket = make_zmq_socket(
-            self.ctx,
-            socket_path,
-            zmq.REP,  # type: ignore[attr-defined]
-            bind=True,
-        )
-
-        self.lmcache_engine = lmcache_engine
-        self.running = True
-
-        def process_request():
-            while self.running:
-                # try:
-                frames = self.socket.recv_multipart(copy=False)
-                user = pickle.loads(frames[0])
-                token_ids = pickle.loads(frames[1])
-                result = self.lmcache_engine.lookup(token_ids, pin=True, user=user)
-                response = result.to_bytes(4, "big")
-                self.socket.send(response)
-                # except Exception as e:
-                #    logger.error("Error in LMCache lookup server: %s", e)
-                #    break
-                # continue
-
-        self.thread = threading.Thread(target=process_request, daemon=True)
-        self.thread.start()
-        logger.info(
-            f"Started LMCache lookup server for tp = {vllm_config.parallel_config.rank}"
-        )
-
-    def close(self):
-        self.socket.close(linger=0)
-        # TODO: close the thread!
-
-
 @dataclass
 class LoadSpec:
     # Number of tokens cached in vLLM
